% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{book_skeleton}
\alias{book_skeleton}
\title{Install book files}
\usage{
book_skeleton(path)
}
\arguments{
\item{path}{The path to intall the ressources to}
}
\value{
TRUE
}
\description{
Install book files
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
book_skeleton("./")
\dontshow{\}) # examplesIf}
}
