% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_classic_box}
\alias{eng_classic_box}
\title{Classic box environment}
\usage{
eng_classic_box(options)
}
\arguments{
\item{options}{}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
