% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_exam_questions}
\alias{eng_exam_questions}
\title{Question in exam document}
\usage{
eng_exam_questions(options)
}
\arguments{
\item{options}{Chunk options}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use quarto extension with solution filter.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: knitr engine for Rmd/Quarto documents
eng_exam_questions(code = "Some text")
\dontshow{\}) # examplesIf}
}
