% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_image_legend}
\alias{eng_image_legend}
\title{Insert figure environment}
\usage{
eng_image_legend(options)
}
\arguments{
\item{options}{Chunk options}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: knitr engine for Rmd/Quarto documents
eng_image_legend(list(
  img.cap = "Image caption",
  img.width = 80,
  img.author = "Someone"
))
\dontshow{\}) # examplesIf}
}
