% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_legal_list}
\alias{eng_legal_list}
\title{Create paralist}
\usage{
eng_legal_list(options)
}
\arguments{
\item{options}{Chunk options}
}
\description{
Create paralist
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: knitr engine for Rmd/Quarto documents
eng_legal_list(code = "Some text")
\dontshow{\}) # examplesIf}
}
