% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_wrap_figure}
\alias{eng_wrap_figure}
\title{Wrap image}
\usage{
eng_wrap_figure(options)
}
\arguments{
\item{options}{Chunk options}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: knitr engine for Rmd/Quarto documents
eng_wrap_figure(list(
  fig.cap = "Image caption",
  out.width = "80\%",
  wrap.width = "20\%"
))
\dontshow{\}) # examplesIf}
}
