% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{ggeo_knit_theme}
\alias{ggeo_knit_theme}
\title{Plot theme for knitted documents}
\usage{
ggeo_knit_theme(..., theme = "doc", mode = "light")
}
\arguments{
\item{theme}{Override \code{\link[ggeo:ggeotheme]{ggeo::ggeotheme()}} theme parameter}

\item{mode}{Override \code{\link[ggeo:ggeotheme]{ggeo::ggeotheme()}} mode parameter}
}
\value{
A ggplot2 theme
}
\description{
Plot theme for knitted documents
}
\seealso{
\code{\link[ggeo:ggeotheme]{ggeo::ggeotheme()}}
}
