% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_book}
\alias{gvz_book}
\title{Book pdf}
\usage{
gvz_book(..., metadata = c())
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=gvz_render_pdf_book]{gvz_render_pdf_book}}
  \describe{
    \item{\code{template_path}}{Path of the latex template}
  }}

\item{metadata}{Additional pandoc metadata}
}
\description{
Book pdf
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: should be used as Rmd format in yaml frontmatter
#
# format: geovizir::gvz_book
\dontshow{\}) # examplesIf}
}
