% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_bs4_book}
\alias{gvz_bs4_book}
\title{Book website}
\usage{
gvz_bs4_book(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bookdown:bs4_book]{bookdown::bs4_book}}
  \describe{
    \item{\code{theme}}{A named list or \code{\link[bslib:bs_theme]{bslib::bs_theme()}} object.
The default, \code{bs4_book_theme()}, resets the base font size to 1rem to
make reading easier and uses a primary colour with greater contrast
against the background.}
    \item{\code{repo}}{Either link to repository where book is hosted, used to generate
view source and edit buttons or a list with repository \code{base} link, default
\code{branch}, \code{subdir} and \code{icon}
(see "Specifying the repository" in \url{https://bookdown.org/yihui/bookdown/html.html#bootstrap4-style}).}
    \item{\code{template}}{Pandoc template to use for rendering. Pass \code{"default"} to use
the bookdown default template; pass a path to use a custom template. The
default template should be sufficient for most use cases. For advanced user
only, in case you want to develop a custom template, we highly recommend to
start from the default template:
\url{https://github.com/rstudio/bookdown/blob/master/inst/templates/bs4_book.html}.
Otherwise, some feature may not work anymore.}
    \item{\code{footnotes_inline}}{By default, footnotes will be set inline and shown on
hover. Set to \code{FALSE} to keep footnotes at the bottom of the page with
links.}
  }}
}
\description{
Book website
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: should be used as Rmd format in yaml frontmatter
#
# format: geovizir::gvz_bs4_book
\dontshow{\}) # examplesIf}
}
