% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_cosmology.R
\name{gvz_cosmo_solar_system_object}
\alias{gvz_cosmo_solar_system_object}
\title{Format cosmology object box}
\usage{
gvz_cosmo_solar_system_object(name)
}
\value{

}
\description{
Format cosmology object box
}
