% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_document}
\alias{gvz_document}
\title{Standard pdf document}
\usage{
gvz_document(..., metadata = c())
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=gvz_render_pdf_document]{gvz_render_pdf_document}}
  \describe{
    \item{\code{template_path}}{Path of the latex template}
  }}

\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf document
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: should be used as Rmd format in yaml frontmatter
#
# format: geovizir::gvz_document
\dontshow{\}) # examplesIf}
}
