% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_html_or_pdf}
\alias{gvz_html_or_pdf}
\title{Select output depending on format}
\usage{
gvz_html_or_pdf(html_output, pdf_output)
}
\arguments{
\item{html_output}{Output for html format}

\item{pdf_output}{Output for pdf format}
}
\value{
The selected output
}
\description{
Select output depending on format
}
