% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{gvz_install_fonts_macos}
\alias{gvz_install_fonts_macos}
\title{Install fonts on MacOS}
\usage{
gvz_install_fonts_macos(system_wide = FALSE)
}
\arguments{
\item{system_wide}{Install the fonts for all users?}
}
\value{
Move fonts to MacOS font folder
}
\description{
Install fonts on MacOS
}
