% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_children.R
\name{gvz_knit_child}
\alias{gvz_knit_child}
\title{Knit child for MF oral exams}
\usage{
gvz_knit_child(input, ...)
}
\arguments{
\item{...}{Data that is passed to the fragment
Must have \emph{ID}, \emph{Name}, \emph{Subject} and \emph{Question}}
}
\description{
Knits the child using \code{cat()} to
print the output.
}
\details{
This function should be used in a Quarto document inside
a code chunk with the \code{asis} option.
}
