% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_children.R
\name{gvz_knit_child_matu_oraux}
\alias{gvz_knit_child_matu_oraux}
\title{Knit child for MF oral exams}
\usage{
gvz_knit_child_matu_oraux(...)
}
\arguments{
\item{...}{}
}
\value{
Knitted document
}
\description{
Knit child for MF oral exams
}
