% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_letter_standard}
\alias{gvz_letter_standard}
\title{Standard pdf letter}
\usage{
gvz_letter_standard(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf letter
}
\keyword{internal}
