% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matu_federale.R
\name{gvz_matu_data}
\alias{gvz_matu_data}
\title{Get exam data for Matu fédérale}
\usage{
gvz_matu_data(start_date, end_date)
}
\arguments{
\item{start_date}{String Y-M-D start date for filter}

\item{end_date}{String Y-M-D end date for filter}
}
\value{
A tibble with exam data
}
\description{
Get exam data for Matu fédérale
}
\keyword{internal}
