% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matu_federale.R
\name{gvz_matu_data_assemble}
\alias{gvz_matu_data_assemble}
\title{Assemble exam data into a tibble}
\usage{
gvz_matu_data_assemble(page)
}
\arguments{
\item{page}{Notion page object}
}
\value{
A tibble
}
\description{
Assemble exam data into a tibble
}
\keyword{internal}
