% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_matu_oraux}
\alias{gvz_matu_oraux}
\title{Oraux matu fédérale}
\usage{
gvz_matu_oraux(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Oraux matu fédérale
}
