% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_md_to_latex}
\alias{gvz_md_to_latex}
\title{Helper to convert from Markdown to LaTeX}
\usage{
gvz_md_to_latex(content)
}
\arguments{
\item{content}{String to convert}
}
\value{
A string in LaTeX
}
\description{
Helper to convert from Markdown to LaTeX
}
