% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_metadata}
\alias{gvz_metadata}
\title{Generate pandoc metadata}
\usage{
gvz_metadata(path)
}
\arguments{
\item{path}{Path of template files}

\item{knit_key}{Boolean: knit key file?}
}
\value{
A vector with pandoc metadata arguments
}
\description{
Generate pandoc metadata
}
\keyword{internal}
