% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{gvz_quarto_get_presentation_yaml}
\alias{gvz_quarto_get_presentation_yaml}
\title{Gets presentation yaml file from GeoKey}
\usage{
gvz_quarto_get_presentation_yaml(path, prefix)
}
\arguments{
\item{path}{Path where the yaml file gets downloaded to}

\item{prefix}{Prefix for the path on GeoKey}
}
\description{
Gets presentation yaml file from GeoKey
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: only work in quarto projects

gvz_quarto_get_presentation_yaml("Presentations/", "DF")
\dontshow{\}) # examplesIf}
}
