% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{gvz_quarto_make_presentation_yaml}
\alias{gvz_quarto_make_presentation_yaml}
\title{Make a simple yaml file for all presentations}
\usage{
gvz_quarto_make_presentation_yaml()
}
\description{
Make a simple yaml file for all presentations
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: only work in quarto projects

gvz_quarto_make_presentation_yaml()
\dontshow{\}) # examplesIf}
}
