% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{gvz_quarto_setup}
\alias{gvz_quarto_setup}
\title{Quarto chunk options setup}
\usage{
gvz_quarto_setup()
}
\description{
Set useful chunk options in Quarto documents.
}
\examples{
gvz_quarto_setup()

}
