% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_render}
\alias{gvz_render}
\title{Render documents}
\usage{
gvz_render(..., template_path, metadata)
}
\arguments{
\item{template_path}{Path of the latex template}

\item{metadata}{Vector with pandoc metadata}
}
\description{
Render documents
}
\keyword{internal}
