% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_render_diagrams}
\alias{gvz_render_diagrams}
\title{Render all tikz diagram in PDF format}
\usage{
gvz_render_diagrams(source, out)
}
\arguments{
\item{source}{Source file with tikz chunks}

\item{out}{Path for diagram's to export to}
}
\description{
Render all tikz diagram in PDF format
}
