% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_render_pdf_document}
\alias{gvz_render_pdf_document}
\title{Render pdf documents}
\usage{
gvz_render_pdf_document(..., template_path, metadata)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[bookdown:html_document2]{bookdown::pdf_document2}}
  \describe{
    \item{\code{}}{}
  }}

\item{template_path}{Path of the latex template}

\item{metadata}{Vector with pandoc metadata}
}
\description{
Render pdf documents
}
\keyword{internal}
