% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_render.R
\name{gvz_render_test_folders}
\alias{gvz_render_test_folders}
\title{Render Test folders}
\usage{
gvz_render_test_folders(students, template, output_dir)
}
\arguments{
\item{students}{A tibble of students}

\item{template}{A quarto template}
}
\value{
Generate pdf files
}
\description{
Render Test folders
}
