% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_tikz_diagram}
\alias{gvz_tikz_diagram}
\title{Generate tikz diagram}
\usage{
gvz_tikz_diagram(label, out)
}
\arguments{
\item{label}{Tikz chunk label}

\item{out}{Path for figure}
}
\description{
Generate tikz diagram
}
\keyword{internal}
