% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_utils.R
\name{gvz_walk_multiple}
\alias{gvz_walk_multiple}
\title{Walk the folders}
\usage{
gvz_walk_multiple(template, output_dir, ...)
}
\arguments{
\item{template}{The template}

\item{...}{The parameter for the rest}
}
\description{
Walk the folders
}
\keyword{internal}
