% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_render.R
\name{gvz_walk_test_folders}
\alias{gvz_walk_test_folders}
\title{Walk the folders}
\usage{
gvz_walk_test_folders(template, output_dir, ...)
}
\arguments{
\item{template}{The template}

\item{...}{The parameter for the rest}
}
\value{
Nothing
}
\description{
Walk the folders
}
\keyword{internal}
