% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{knit_cleanup}
\alias{knit_cleanup}
\title{Remove all temporary files}
\usage{
knit_cleanup(input)
}
\arguments{
\item{input}{The path to clean}
}
\description{
Remove all temporary files
}
\keyword{internal}
