% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{knit_quiet}
\alias{knit_quiet}
\title{Custom Knit function for RStudio}
\usage{
knit_quiet(input, ...)
}
\arguments{
\item{input}{Input file}

\item{...}{Unused, for compatibility in yaml frontmatter}
}
\description{
Custom Knit function for RStudio
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: this function is for use in Rmd yaml frontmatter
#
# knit: geovizir::knit_quiet
\dontshow{\}) # examplesIf}
}
