% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{knit_subfile}
\alias{knit_subfile}
\title{Knit subfiles}
\usage{
knit_subfile(input, output, key, value)
}
\arguments{
\item{input}{The input file}

\item{output}{The output file}

\item{key}{Key to add to output filename}

\item{value}{Subclass value}
}
\description{
Knit subfiles
}
\keyword{internal}
