% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{latex_percent_size}
\alias{latex_percent_size}
\title{Return latex sizes}
\usage{
latex_percent_size(x, which = c("width", "height"), width = "\\\\linewidth")
}
\arguments{
\item{x}{The size in percents}

\item{which}{One of \code{width} or \code{height}}

\item{width}{The width measure to use}
}
\value{
Latex size
}
\description{
Return latex sizes
}
\keyword{internal}
