% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{viz_circle}
\alias{viz_circle}
\title{Circle layer}
\usage{
viz_circle(
  map,
  data = NULL,
  id = NULL,
  pos = c(0, 0),
  r = 10,
  k = 50,
  fixmax = NULL,
  dodge = FALSE,
  iteration = 200,
  sort = NULL,
  descending = NULL,
  coords = "geo",
  fill = NULL,
  stroke = "white",
  tip = FALSE,
  ...
)
}
\arguments{
\item{map}{A \code{geovizr} map object created using \code{viz_create()}.}

\item{data}{object. Optional. A spatial data object, such as an `sf` object or a `terra::SpatVector`.}

\item{id}{character. Optional. Unique layer id.}

\item{pos}{numeric vector. Optional. Position of a single circle (default c(0, 0)).}

\item{r}{numeric or character. Optional. Circle radius (default 10).
Can be a fixed value or the name of a field containing numerical values.}

\item{k}{numeric. Optional. Radius of the largest circle (default 50).}

\item{fixmax}{numeric. Optional. Value corresponding to the circle of radius \code{k}.
Useful to ensure comparability between maps.}

\item{dodge}{logical. Optional. Avoid circle overlap (default FALSE).}

\item{iteration}{numeric. Optional. Number of iterations for dodging (default 200).}

\item{sort}{character or function. Optional. Field name or function to sort circles.}

\item{descending}{logical. Optional. Sorting order.}

\item{coords}{character. Optional. Coordinate system (default "geo").
Use \code{"svg"} if coordinates are already expressed in the SVG coordinate space.}

\item{fill}{character or function. Optional. Fill color.}

\item{stroke}{character or function. Optional. Stroke color (default "white").}

\item{tip}{logical or function. Optional. Tooltip definition (default FALSE).
Use TRUE to display all fields.}

\item{...}{Additional SVG attributes (e.g. \code{strokeDasharray}, \code{strokeWidth},
\code{opacity}, \code{strokeLinecap}, etc.).}
}
\value{
A modified `geoviz` map object with a new layer added.
Rendering is performed using \code{viz_render()}.
}
\description{
The \code{viz_circle} function draws circles on the map from a spatial
data frame or from a single position. It can be used to create proportional symbol maps
with optional collision avoidance.
}
\examples{
library(sf)
world <- st_read(
  system.file("gpkg/world.gpkg", package = "geovizr"),
  quiet = TRUE
)

viz_create(projection = "EqualEarth") |>
  viz_path(datum = world, fill = "#f1f3f5") |>
  viz_circle(data = world, r = 30, fill = "#38896F") |>
  viz_render()
}
