% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\name{viz_typo}
\alias{viz_typo}
\title{Typology layer}
\usage{
viz_typo(
  map,
  data = NULL,
  var,
  colors = NULL,
  order = NULL,
  alphabetical = TRUE,
  missing = "white",
  legend = TRUE,
  leg_type = "vertical",
  leg_pos = c(10, 10),
  ...
)
}
\arguments{
\item{map}{A \code{geovizr} map object created using \code{viz_create()}.}

\item{data}{A spatial data object, such as an `sf` object or a `terra::SpatVector`. Use \code{data} to enable iteration over features.}

\item{var}{character. Name of the categorical variable used for styling.
You can also use \code{fill} or \code{stroke} directly instead of \code{var}.}

\item{colors}{character or vector. Optional. Color palette or vector of colors.
Can use named palettes from the Dicopal library (e.g. "Antique", "Bold", "Pastel",
"Prism", "Safe", "Vivid", "Accent", "Dark2", "Paired", "Pastel1", "Pastel2",
"Set1", "Set2", "Set3"). See https://observablehq.com/@neocartocnrs/dicopal-library}

\item{order}{character vector. Optional. Explicit ordering of categories.}

\item{alphabetical}{logical. Optional. Whether to sort legend items alphabetically
(default TRUE).}

\item{missing}{character or logical. Optional. Color for missing values (default "white").}

\item{legend}{logical. Optional. Whether to display a legend (default TRUE).}

\item{leg_type}{character. Optional. Legend orientation (default "vertical").
One of: "horizontal", "vertical".}

\item{leg_pos}{numeric vector. Optional. Legend position (default c(10, 10)).}

\item{...}{Additional parameters passed to path rendering. With the \code{leg_} prefix, you can configure the legend. For example: \code{leg_title}, \code{leg_subtitle}, \code{leg_note}, etc.}
}
\value{
A modified `geoviz` map object with a new layer added.
Rendering is performed using \code{viz_render()}.
}
\description{
The \code{viz_typo} function creates a typology map from a spatial
data frame by mapping categorical variables to colors. It supports custom ordering,
color palettes, and automatic legend generation.
}
\examples{
library(sf)
world <- st_read(
  system.file("gpkg/world.gpkg", package = "geovizr"),
  quiet = TRUE
)

viz_create(
  projection = "EqualEarth", background = "white",
  zoomable = TRUE
) |>
  viz_typo(
    data = world, var = "region", colors = "Pastel",
    leg_title = "Continents"
  ) |>
  viz_render()
}
