% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{chapter_image}
\alias{chapter_image}
\title{Set chapter image in for LaTeX book}
\usage{
chapter_image(path)
}
\arguments{
\item{path}{The path to the image to use}
}
\description{
Set chapter image in for LaTeX book
}
