% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_engines.R
\name{eng_document_ref}
\alias{eng_document_ref}
\title{Provide a knit engine for AlXxx LaTeX commands}
\usage{
eng_document_ref(options)
}
\arguments{
\item{options}{knitr engine options}
}
\value{
A string
}
\description{
Provide a knit engine for AlXxx LaTeX commands
}
