% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_book_resources}
\alias{gvz_book_resources}
\title{Get the path for the book resources}
\usage{
gvz_book_resources()
}
\value{
A full path with the book resources, defaults to project root
}
\description{
Get the path for the book resources
}
