% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{gvz_doc_theme}
\alias{gvz_doc_theme}
\title{Document ggeo theme}
\usage{
gvz_doc_theme(...)
}
\value{
A ggplot2 theme
}
\description{
Document ggeo theme
}
\seealso{
\code{\link[ggeo:ggeotheme]{ggeo::ggeotheme()}}
}
