% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_document}
\alias{gvz_document}
\title{Standard pdf document}
\usage{
gvz_document(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Standard pdf document
}
