% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gvz_file}
\alias{gvz_file}
\title{A shortcut function for the system.file to the geovizr package}
\usage{
gvz_file(...)
}
\value{
A path
}
\description{
A shortcut function for the system.file to the geovizr package
}
