% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_utils.R
\name{gvz_latex_table}
\alias{gvz_latex_table}
\title{Generate latex code for table from tibble}
\usage{
gvz_latex_table(tbl, md_cols = c())
}
\arguments{
\item{tbl}{A tibble}

\item{md_cols}{Columns to treat as markdown}
}
\value{
A collapsed string to use \code{asis}
}
\description{
Generate latex code for table from tibble
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: use in Rmd/Quarto document
table |>
  gvz_latex_table()
\dontshow{\}) # examplesIf}
}
