% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_matu_oraux}
\alias{gvz_matu_oraux}
\title{Oraux matu fédérale}
\usage{
gvz_matu_oraux(..., metadata = c())
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=gvz_render_pdf_document]{gvz_render_pdf_document}}
  \describe{
    \item{\code{template_path}}{Path of the latex template}
  }}

\item{metadata}{Additional pandoc metadata}
}
\description{
Oraux matu fédérale
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Not run: should be used as Rmd format in yaml frontmatter
#
# format: geovizir::gvz_matu_oraux
\dontshow{\}) # examplesIf}
}
