% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_render.R
\name{gvz_render_multiple}
\alias{gvz_render_multiple}
\title{Render multiple files}
\usage{
gvz_render_multiple(data, template, output_dir, merge = F)
}
\arguments{
\item{template}{A quarto template}

\item{students}{A tibble of values for each file}
}
\value{
Generate pdf files
}
\description{
Render multiple files
}
