% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_render_pdf_book}
\alias{gvz_render_pdf_book}
\title{Render pdf books}
\usage{
gvz_render_pdf_book(..., template_path, metadata)
}
\arguments{
\item{template_path}{Path of the latex template}

\item{metadata}{Vector with pandoc metadata}
}
\description{
Render pdf books
}
\keyword{internal}
