% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.R
\name{gvz_reveal_theme}
\alias{gvz_reveal_theme}
\title{Theme function for reveal presentations}
\usage{
gvz_reveal_theme(...)
}
\arguments{
\item{...}{See \code{\link[ggeo:ggeotheme]{ggeo::ggeotheme()}}}
}
\value{
An object of class \code{\link[ggplot2:theme]{ggplot2::theme()}}.
}
\description{
Theme function for reveal presentations
}
