% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_outputs.R
\name{gvz_test_folder}
\alias{gvz_test_folder}
\title{Test with cover page pdf}
\usage{
gvz_test_folder(..., metadata = c())
}
\arguments{
\item{metadata}{Additional pandoc metadata}
}
\description{
Test with cover page pdf
}
